DynamoDB: Sort results of scan/query, fixing the pagination
