+++ tests/test_dynamodb/test_dynamodb.py
@@ -1313,379 +1313,6 @@ def test_filter_expression():
 
 
 @mock_aws
-def test_query_filter():
-    client = boto3.client("dynamodb", region_name="us-east-1")
-    dynamodb = boto3.resource("dynamodb", region_name="us-east-1")
-
-    # Create the DynamoDB table.
-    client.create_table(
-        TableName="test1",
-        AttributeDefinitions=[
-            {"AttributeName": "client", "AttributeType": "S"},
-            {"AttributeName": "app", "AttributeType": "S"},
-        ],
-        KeySchema=[
-            {"AttributeName": "client", "KeyType": "HASH"},
-            {"AttributeName": "app", "KeyType": "RANGE"},
-        ],
-        ProvisionedThroughput={"ReadCapacityUnits": 123, "WriteCapacityUnits": 123},
-    )
-    client.put_item(
-        TableName="test1",
-        Item={
-            "client": {"S": "client1"},
-            "app": {"S": "app1"},
-            "nested": {
-                "M": {
-                    "version": {"S": "version1"},
-                    "contents": {"L": [{"S": "value1"}, {"S": "value2"}]},
-                }
-            },
-        },
-    )
-    client.put_item(
-        TableName="test1",
-        Item={
-            "client": {"S": "client1"},
-            "app": {"S": "app2"},
-            "nested": {
-                "M": {
-                    "version": {"S": "version2"},
-                    "contents": {"L": [{"S": "value1"}, {"S": "value2"}]},
-                }
-            },
-        },
-    )
-
-    table = dynamodb.Table("test1")
-    response = table.query(KeyConditionExpression=Key("client").eq("client1"))
-    assert response["Count"] == 2
-    assert response["ScannedCount"] == 2
-
-    response = table.query(
-        KeyConditionExpression=Key("client").eq("client1"),
-        FilterExpression=Attr("app").eq("app2"),
-    )
-    assert response["Count"] == 1
-    assert response["ScannedCount"] == 2
-    assert response["Items"][0]["app"] == "app2"
-    response = table.query(
-        KeyConditionExpression=Key("client").eq("client1"),
-        FilterExpression=Attr("app").contains("app"),
-    )
-    assert response["Count"] == 2
-
-    response = table.query(
-        KeyConditionExpression=Key("client").eq("client1"),
-        FilterExpression=Attr("nested.version").contains("version"),
-    )
-    assert response["Count"] == 2
-
-    response = table.query(
-        KeyConditionExpression=Key("client").eq("client1"),
-        FilterExpression=Attr("nested.contents[0]").eq("value1"),
-    )
-    assert response["Count"] == 2
-
-    # Combine Limit + Scan
-    response = table.query(
-        KeyConditionExpression=Key("client").eq("client1"),
-        Limit=1,
-        ScanIndexForward=False,
-    )
-    assert response["Count"] == 1
-    assert response["ScannedCount"] == 1
-    assert response["Items"][0]["app"] == "app2"
-
-    response = table.query(
-        KeyConditionExpression=Key("client").eq("client1"),
-        Limit=1,
-        ScanIndexForward=True,
-    )
-    assert response["Count"] == 1
-    assert response["ScannedCount"] == 1
-    assert response["Items"][0]["app"] == "app1"
-
-
-@mock_aws
-def test_query_filter_overlapping_expression_prefixes():
-    client = boto3.client("dynamodb", region_name="us-east-1")
-    dynamodb = boto3.resource("dynamodb", region_name="us-east-1")
-
-    # Create the DynamoDB table.
-    client.create_table(
-        TableName="test1",
-        AttributeDefinitions=[
-            {"AttributeName": "client", "AttributeType": "S"},
-            {"AttributeName": "app", "AttributeType": "S"},
-        ],
-        KeySchema=[
-            {"AttributeName": "client", "KeyType": "HASH"},
-            {"AttributeName": "app", "KeyType": "RANGE"},
-        ],
-        ProvisionedThroughput={"ReadCapacityUnits": 123, "WriteCapacityUnits": 123},
-    )
-
-    client.put_item(
-        TableName="test1",
-        Item={
-            "client": {"S": "client1"},
-            "app": {"S": "app1"},
-            "nested": {
-                "M": {
-                    "version": {"S": "version1"},
-                    "contents": {"L": [{"S": "value1"}, {"S": "value2"}]},
-                }
-            },
-        },
-    )
-
-    table = dynamodb.Table("test1")
-    response = table.query(
-        KeyConditionExpression=Key("client").eq("client1") & Key("app").eq("app1"),
-        ProjectionExpression="#1, #10, nested",
-        ExpressionAttributeNames={"#1": "client", "#10": "app"},
-    )
-
-    assert response["Count"] == 1
-    assert response["Items"][0] == {
-        "client": "client1",
-        "app": "app1",
-        "nested": {"version": "version1", "contents": ["value1", "value2"]},
-    }
-
-
-@mock_aws
-def test_scan_filter():
-    client = boto3.client("dynamodb", region_name="us-east-1")
-    dynamodb = boto3.resource("dynamodb", region_name="us-east-1")
-
-    # Create the DynamoDB table.
-    client.create_table(
-        TableName="test1",
-        AttributeDefinitions=[
-            {"AttributeName": "client", "AttributeType": "S"},
-            {"AttributeName": "app", "AttributeType": "S"},
-        ],
-        KeySchema=[
-            {"AttributeName": "client", "KeyType": "HASH"},
-            {"AttributeName": "app", "KeyType": "RANGE"},
-        ],
-        ProvisionedThroughput={"ReadCapacityUnits": 123, "WriteCapacityUnits": 123},
-    )
-    client.put_item(
-        TableName="test1", Item={"client": {"S": "client1"}, "app": {"S": "app1"}}
-    )
-
-    table = dynamodb.Table("test1")
-    response = table.scan(FilterExpression=Attr("app").eq("app2"))
-    assert response["Count"] == 0
-
-    response = table.scan(FilterExpression=Attr("app").eq("app1"))
-    assert response["Count"] == 1
-
-    response = table.scan(FilterExpression=Attr("app").ne("app2"))
-    assert response["Count"] == 1
-
-    response = table.scan(FilterExpression=Attr("app").ne("app1"))
-    assert response["Count"] == 0
-
-
-@mock_aws
-def test_scan_filter2():
-    client = boto3.client("dynamodb", region_name="us-east-1")
-
-    # Create the DynamoDB table.
-    client.create_table(
-        TableName="test1",
-        AttributeDefinitions=[
-            {"AttributeName": "client", "AttributeType": "S"},
-            {"AttributeName": "app", "AttributeType": "N"},
-        ],
-        KeySchema=[
-            {"AttributeName": "client", "KeyType": "HASH"},
-            {"AttributeName": "app", "KeyType": "RANGE"},
-        ],
-        ProvisionedThroughput={"ReadCapacityUnits": 123, "WriteCapacityUnits": 123},
-    )
-    client.put_item(
-        TableName="test1", Item={"client": {"S": "client1"}, "app": {"N": "1"}}
-    )
-
-    response = client.scan(
-        TableName="test1",
-        Select="ALL_ATTRIBUTES",
-        FilterExpression="#tb >= :dt",
-        ExpressionAttributeNames={"#tb": "app"},
-        ExpressionAttributeValues={":dt": {"N": str(1)}},
-    )
-    assert response["Count"] == 1
-
-
-@mock_aws
-def test_scan_filter3():
-    client = boto3.client("dynamodb", region_name="us-east-1")
-    dynamodb = boto3.resource("dynamodb", region_name="us-east-1")
-
-    # Create the DynamoDB table.
-    client.create_table(
-        TableName="test1",
-        AttributeDefinitions=[
-            {"AttributeName": "client", "AttributeType": "S"},
-            {"AttributeName": "app", "AttributeType": "N"},
-        ],
-        KeySchema=[
-            {"AttributeName": "client", "KeyType": "HASH"},
-            {"AttributeName": "app", "KeyType": "RANGE"},
-        ],
-        ProvisionedThroughput={"ReadCapacityUnits": 123, "WriteCapacityUnits": 123},
-    )
-    client.put_item(
-        TableName="test1",
-        Item={"client": {"S": "client1"}, "app": {"N": "1"}, "active": {"BOOL": True}},
-    )
-
-    table = dynamodb.Table("test1")
-    response = table.scan(FilterExpression=Attr("active").eq(True))
-    assert response["Count"] == 1
-
-    response = table.scan(FilterExpression=Attr("active").ne(True))
-    assert response["Count"] == 0
-
-    response = table.scan(FilterExpression=Attr("active").ne(False))
-    assert response["Count"] == 1
-
-    response = table.scan(FilterExpression=Attr("app").ne(1))
-    assert response["Count"] == 0
-
-    response = table.scan(FilterExpression=Attr("app").ne(2))
-    assert response["Count"] == 1
-
-
-@mock_aws
-def test_scan_filter4():
-    client = boto3.client("dynamodb", region_name="us-east-1")
-    dynamodb = boto3.resource("dynamodb", region_name="us-east-1")
-
-    # Create the DynamoDB table.
-    client.create_table(
-        TableName="test1",
-        AttributeDefinitions=[
-            {"AttributeName": "client", "AttributeType": "S"},
-            {"AttributeName": "app", "AttributeType": "N"},
-        ],
-        KeySchema=[
-            {"AttributeName": "client", "KeyType": "HASH"},
-            {"AttributeName": "app", "KeyType": "RANGE"},
-        ],
-        ProvisionedThroughput={"ReadCapacityUnits": 123, "WriteCapacityUnits": 123},
-    )
-
-    table = dynamodb.Table("test1")
-    response = table.scan(
-        FilterExpression=Attr("epoch_ts").lt(7) & Attr("fanout_ts").not_exists()
-    )
-    # Just testing
-    assert response["Count"] == 0
-
-
-@mock_aws
-def test_scan_filter_should_not_return_non_existing_attributes():
-    table_name = "my-table"
-    item = {"partitionKey": "pk-2", "my-attr": 42}
-    # Create table
-    res = boto3.resource("dynamodb", region_name="us-east-1")
-    res.create_table(
-        TableName=table_name,
-        KeySchema=[{"AttributeName": "partitionKey", "KeyType": "HASH"}],
-        AttributeDefinitions=[{"AttributeName": "partitionKey", "AttributeType": "S"}],
-        BillingMode="PAY_PER_REQUEST",
-    )
-    table = res.Table(table_name)
-    # Insert items
-    table.put_item(Item={"partitionKey": "pk-1"})
-    table.put_item(Item=item)
-    # Verify a few operations
-    # Assert we only find the item that has this attribute
-    assert table.scan(FilterExpression=Attr("my-attr").lt(43))["Items"] == [item]
-    assert table.scan(FilterExpression=Attr("my-attr").lte(42))["Items"] == [item]
-    assert table.scan(FilterExpression=Attr("my-attr").gte(42))["Items"] == [item]
-    assert table.scan(FilterExpression=Attr("my-attr").gt(41))["Items"] == [item]
-    # Sanity check that we can't find the item if the FE is wrong
-    assert table.scan(FilterExpression=Attr("my-attr").gt(43))["Items"] == []
-
-
-@mock_aws
-def test_bad_scan_filter():
-    client = boto3.client("dynamodb", region_name="us-east-1")
-    dynamodb = boto3.resource("dynamodb", region_name="us-east-1")
-
-    # Create the DynamoDB table.
-    client.create_table(
-        TableName="test1",
-        AttributeDefinitions=[
-            {"AttributeName": "client", "AttributeType": "S"},
-            {"AttributeName": "app", "AttributeType": "S"},
-        ],
-        KeySchema=[
-            {"AttributeName": "client", "KeyType": "HASH"},
-            {"AttributeName": "app", "KeyType": "RANGE"},
-        ],
-        ProvisionedThroughput={"ReadCapacityUnits": 123, "WriteCapacityUnits": 123},
-    )
-    table = dynamodb.Table("test1")
-
-    # Bad expression
-    with pytest.raises(ClientError) as exc:
-        table.scan(FilterExpression="client test")
-    assert exc.value.response["Error"]["Code"] == "ValidationException"
-
-
-@mock_aws
-def test_scan_with_scanfilter():
-    table_name = "my-table"
-    item = {"partitionKey": "pk-2", "my-attr": 42}
-    client = boto3.client("dynamodb", region_name="us-east-1")
-    res = boto3.resource("dynamodb", region_name="us-east-1")
-    res.create_table(
-        TableName=table_name,
-        KeySchema=[{"AttributeName": "partitionKey", "KeyType": "HASH"}],
-        AttributeDefinitions=[{"AttributeName": "partitionKey", "AttributeType": "S"}],
-        BillingMode="PAY_PER_REQUEST",
-    )
-    table = res.Table(table_name)
-    table.put_item(Item={"partitionKey": "pk-1"})
-    table.put_item(Item=item)
-
-    # ScanFilter: EQ
-    # The DynamoDB table-resource sends the AttributeValueList in the wrong format
-    # So this operation never finds any data, in Moto or AWS
-    table.scan(
-        ScanFilter={
-            "my-attr": {"AttributeValueList": [{"N": "42"}], "ComparisonOperator": "EQ"}
-        }
-    )
-
-    # ScanFilter: EQ
-    # If we use the boto3-client, we do receive the correct data
-    items = client.scan(
-        TableName=table_name,
-        ScanFilter={
-            "partitionKey": {
-                "AttributeValueList": [{"S": "pk-1"}],
-                "ComparisonOperator": "EQ",
-            }
-        },
-    )["Items"]
-    assert items == [{"partitionKey": {"S": "pk-1"}}]
-
-    # ScanFilter: NONE
-    # Note that we can use the table-resource here, because we're not using the AttributeValueList
-    items = table.scan(ScanFilter={"my-attr": {"ComparisonOperator": "NULL"}})["Items"]
-    assert items == [{"partitionKey": "pk-1"}]
-
-
-@mock_aws
 def test_duplicate_create():
     client = boto3.client("dynamodb", region_name="us-east-1")
 
@@ -2410,63 +2037,6 @@ def test_query_global_secondary_index_when_created_via_update_table_resource():
 
 
 @mock_aws
-def test_query_gsi_with_range_key():
-    dynamodb = boto3.client("dynamodb", region_name="us-east-1")
-    dynamodb.create_table(
-        TableName="test",
-        KeySchema=[{"AttributeName": "id", "KeyType": "HASH"}],
-        AttributeDefinitions=[
-            {"AttributeName": "id", "AttributeType": "S"},
-            {"AttributeName": "gsi_hash_key", "AttributeType": "S"},
-            {"AttributeName": "gsi_range_key", "AttributeType": "S"},
-        ],
-        ProvisionedThroughput={"ReadCapacityUnits": 1, "WriteCapacityUnits": 1},
-        GlobalSecondaryIndexes=[
-            {
-                "IndexName": "test_gsi",
-                "KeySchema": [
-                    {"AttributeName": "gsi_hash_key", "KeyType": "HASH"},
-                    {"AttributeName": "gsi_range_key", "KeyType": "RANGE"},
-                ],
-                "Projection": {"ProjectionType": "ALL"},
-                "ProvisionedThroughput": {
-                    "ReadCapacityUnits": 1,
-                    "WriteCapacityUnits": 1,
-                },
-            }
-        ],
-    )
-
-    dynamodb.put_item(
-        TableName="test",
-        Item={
-            "id": {"S": "test1"},
-            "gsi_hash_key": {"S": "key1"},
-            "gsi_range_key": {"S": "range1"},
-        },
-    )
-    dynamodb.put_item(
-        TableName="test", Item={"id": {"S": "test2"}, "gsi_hash_key": {"S": "key1"}}
-    )
-
-    res = dynamodb.query(
-        TableName="test",
-        IndexName="test_gsi",
-        KeyConditionExpression="gsi_hash_key = :gsi_hash_key and gsi_range_key = :gsi_range_key",
-        ExpressionAttributeValues={
-            ":gsi_hash_key": {"S": "key1"},
-            ":gsi_range_key": {"S": "range1"},
-        },
-    )
-    assert res["Count"] == 1
-    assert res["Items"][0] == {
-        "id": {"S": "test1"},
-        "gsi_hash_key": {"S": "key1"},
-        "gsi_range_key": {"S": "range1"},
-    }
-
-
-@mock_aws
 def test_scan_by_non_exists_index():
     dynamodb = boto3.client("dynamodb", region_name="us-east-1")
 
@@ -2919,43 +2489,6 @@ def create_table_with_list(table_name):
         BillingMode="PAY_PER_REQUEST",
     )
     return client
-
-
-@mock_aws
-def test_sorted_query_with_numerical_sort_key():
-    dynamodb = boto3.resource("dynamodb", region_name="us-east-1")
-    dynamodb.create_table(
-        TableName="CarCollection",
-        KeySchema=[
-            {"AttributeName": "CarModel", "KeyType": "HASH"},
-            {"AttributeName": "CarPrice", "KeyType": "RANGE"},
-        ],
-        AttributeDefinitions=[
-            {"AttributeName": "CarModel", "AttributeType": "S"},
-            {"AttributeName": "CarPrice", "AttributeType": "N"},
-        ],
-        ProvisionedThroughput={"ReadCapacityUnits": 1, "WriteCapacityUnits": 1},
-    )
-
-    def create_item(price):
-        return {"CarModel": "M", "CarPrice": price}
-
-    table = dynamodb.Table("CarCollection")
-    items = list(map(create_item, [2, 1, 10, 3]))
-    for item in items:
-        table.put_item(Item=item)
-
-    response = table.query(KeyConditionExpression=Key("CarModel").eq("M"))
-
-    response_items = response["Items"]
-    assert len(items) == len(response_items)
-    assert all(isinstance(item["CarPrice"], Decimal) for item in response_items)
-    response_prices = [item["CarPrice"] for item in response_items]
-    expected_prices = [Decimal(item["CarPrice"]) for item in items]
-    expected_prices.sort()
-    assert (
-        expected_prices == response_prices
-    ), "result items are not sorted by numerical value"
 
 
 # https://github.com/getmoto/moto/issues/1874
@@ -3674,61 +3207,6 @@ def test_valid_transact_get_items():
 
 
 @mock_aws
-def test_gsi_verify_negative_number_order():
-    table_schema = {
-        "KeySchema": [{"AttributeName": "partitionKey", "KeyType": "HASH"}],
-        "GlobalSecondaryIndexes": [
-            {
-                "IndexName": "GSI-K1",
-                "KeySchema": [
-                    {"AttributeName": "gsiK1PartitionKey", "KeyType": "HASH"},
-                    {"AttributeName": "gsiK1SortKey", "KeyType": "RANGE"},
-                ],
-                "Projection": {"ProjectionType": "KEYS_ONLY"},
-            }
-        ],
-        "AttributeDefinitions": [
-            {"AttributeName": "partitionKey", "AttributeType": "S"},
-            {"AttributeName": "gsiK1PartitionKey", "AttributeType": "S"},
-            {"AttributeName": "gsiK1SortKey", "AttributeType": "N"},
-        ],
-    }
-
-    item1 = {
-        "partitionKey": "pk-1",
-        "gsiK1PartitionKey": "gsi-k1",
-        "gsiK1SortKey": Decimal("-0.6"),
-    }
-
-    item2 = {
-        "partitionKey": "pk-2",
-        "gsiK1PartitionKey": "gsi-k1",
-        "gsiK1SortKey": Decimal("-0.7"),
-    }
-
-    item3 = {
-        "partitionKey": "pk-3",
-        "gsiK1PartitionKey": "gsi-k1",
-        "gsiK1SortKey": Decimal("0.7"),
-    }
-
-    dynamodb = boto3.resource("dynamodb", region_name="us-east-1")
-    dynamodb.create_table(
-        TableName="test-table", BillingMode="PAY_PER_REQUEST", **table_schema
-    )
-    table = dynamodb.Table("test-table")
-    table.put_item(Item=item3)
-    table.put_item(Item=item1)
-    table.put_item(Item=item2)
-
-    resp = table.query(
-        KeyConditionExpression=Key("gsiK1PartitionKey").eq("gsi-k1"), IndexName="GSI-K1"
-    )
-    # Items should be ordered with the lowest number first
-    assert [float(item["gsiK1SortKey"]) for item in resp["Items"]] == [-0.7, -0.6, 0.7]
-
-
-@mock_aws
 def test_transact_write_items_put():
     table_schema = {
         "KeySchema": [{"AttributeName": "id", "KeyType": "HASH"}],
@@ -3887,8 +3365,8 @@ def test_transact_write_items_conditioncheck_passes():
     # Assert all are present
     items = dynamodb.scan(TableName="test-table")["Items"]
     assert len(items) == 2
-    assert items[0] == {"email_address": {"S": "foo@moto.com"}, "id": {"S": "foo"}}
-    assert items[1] == {"email_address": {"S": "bar@moto.com"}, "id": {"S": "bar"}}
+    assert {"email_address": {"S": "foo@moto.com"}, "id": {"S": "foo"}} in items
+    assert {"email_address": {"S": "bar@moto.com"}, "id": {"S": "bar"}} in items
 
 
 @mock_aws
@@ -6013,67 +5491,6 @@ def test_query_with_unknown_last_evaluated_key(table_name=None):
     p4 = client.query(
         TableName=table_name,
         KeyConditionExpression="#h = :h",
-        ExpressionAttributeNames={"#h": "pk"},
-        ExpressionAttributeValues={":h": {"S": "hash_value"}},
-        Limit=1,
-        ExclusiveStartKey=different_key,
-    )
-    assert p4["Items"] == [{"pk": {"S": "hash_value"}, "sk": {"S": "range_value6"}}]
-
-
-@pytest.mark.aws_verified
-@dynamodb_aws_verified(add_range=True)
-def test_scan_with_unknown_last_evaluated_key(table_name=None):
-    client = boto3.client("dynamodb", region_name="us-east-1")
-
-    for i in range(10):
-        client.put_item(
-            TableName=table_name,
-            Item={
-                "pk": {"S": "hash_value"},
-                "sk": {"S": f"range_value{i}"},
-            },
-        )
-
-    p1 = client.scan(
-        TableName=table_name,
-        FilterExpression="#h = :h",
-        ExpressionAttributeNames={"#h": "pk"},
-        ExpressionAttributeValues={":h": {"S": "hash_value"}},
-        Limit=1,
-    )
-    assert p1["Items"] == [{"pk": {"S": "hash_value"}, "sk": {"S": "range_value0"}}]
-
-    # Using the Exact ExclusiveStartKey provided
-    p2 = client.scan(
-        TableName=table_name,
-        FilterExpression="#h = :h",
-        ExpressionAttributeNames={"#h": "pk"},
-        ExpressionAttributeValues={":h": {"S": "hash_value"}},
-        Limit=1,
-        ExclusiveStartKey=p1["LastEvaluatedKey"],
-    )
-    assert p2["Items"] == [{"pk": {"S": "hash_value"}, "sk": {"S": "range_value1"}}]
-
-    # We can change ExclusiveStartKey
-    # It doesn't need to match - it just needs to be >= page1, but < page1
-    different_key = copy.copy(p1["LastEvaluatedKey"])
-    different_key["sk"]["S"] = different_key["sk"]["S"] + "0"
-    p3 = client.scan(
-        TableName=table_name,
-        FilterExpression="#h = :h",
-        ExpressionAttributeNames={"#h": "pk"},
-        ExpressionAttributeValues={":h": {"S": "hash_value"}},
-        Limit=1,
-        ExclusiveStartKey=different_key,
-    )
-    assert p3["Items"] == [{"pk": {"S": "hash_value"}, "sk": {"S": "range_value1"}}]
-
-    # Sanity check - increasing the sk to something much greater will result in a different outcome
-    different_key["sk"]["S"] = "range_value500"
-    p4 = client.scan(
-        TableName=table_name,
-        FilterExpression="#h = :h",
         ExpressionAttributeNames={"#h": "pk"},
         ExpressionAttributeValues={":h": {"S": "hash_value"}},
         Limit=1,
