+++ tests/test_dynamodb/test_dynamodb_table_with_range_key.py
@@ -78,80 +78,6 @@ def test_query_filter_boto3():
 
 
 @mock_aws
-def test_boto3_conditions():
-    dynamodb = boto3.resource("dynamodb", region_name="us-east-1")
-
-    # Create the DynamoDB table.
-    table = dynamodb.create_table(
-        TableName="users",
-        KeySchema=[
-            {"AttributeName": "forum_name", "KeyType": "HASH"},
-            {"AttributeName": "subject", "KeyType": "RANGE"},
-        ],
-        AttributeDefinitions=[
-            {"AttributeName": "forum_name", "AttributeType": "S"},
-            {"AttributeName": "subject", "AttributeType": "S"},
-        ],
-        ProvisionedThroughput={"ReadCapacityUnits": 5, "WriteCapacityUnits": 5},
-    )
-    table = dynamodb.Table("users")
-
-    table.put_item(Item={"forum_name": "the-key", "subject": "123"})
-    table.put_item(Item={"forum_name": "the-key", "subject": "456"})
-    table.put_item(Item={"forum_name": "the-key", "subject": "789"})
-
-    # Test a query returning all items
-    results = table.query(
-        KeyConditionExpression=Key("forum_name").eq("the-key") & Key("subject").gt("1"),
-        ScanIndexForward=True,
-    )
-    expected = ["123", "456", "789"]
-    for index, item in enumerate(results["Items"]):
-        assert item["subject"] == expected[index]
-
-    # Return all items again, but in reverse
-    results = table.query(
-        KeyConditionExpression=Key("forum_name").eq("the-key") & Key("subject").gt("1"),
-        ScanIndexForward=False,
-    )
-    for index, item in enumerate(reversed(results["Items"])):
-        assert item["subject"] == expected[index]
-
-    # Filter the subjects to only return some of the results
-    results = table.query(
-        KeyConditionExpression=Key("forum_name").eq("the-key")
-        & Key("subject").gt("234"),
-        ConsistentRead=True,
-    )
-    assert results["Count"] == 2
-
-    # Filter to return no results
-    results = table.query(
-        KeyConditionExpression=Key("forum_name").eq("the-key")
-        & Key("subject").gt("9999")
-    )
-    assert results["Count"] == 0
-
-    results = table.query(
-        KeyConditionExpression=Key("forum_name").eq("the-key")
-        & Key("subject").begins_with("12")
-    )
-    assert results["Count"] == 1
-
-    results = table.query(
-        KeyConditionExpression=Key("subject").begins_with("7")
-        & Key("forum_name").eq("the-key")
-    )
-    assert results["Count"] == 1
-
-    results = table.query(
-        KeyConditionExpression=Key("forum_name").eq("the-key")
-        & Key("subject").between("567", "890")
-    )
-    assert results["Count"] == 1
-
-
-@mock_aws
 def test_boto3_conditions_ignorecase():
     dynamodb = boto3.client("dynamodb", region_name="us-east-1")
 
@@ -744,107 +670,11 @@ def test_update_item_delete_with_expression():
 
 
 @mock_aws
-def test_boto3_query_gsi_range_comparison():
-    table = _create_table_with_range_key()
-
-    table.put_item(
-        Item={
-            "forum_name": "the-key",
-            "subject": "123",
-            "username": "johndoe",
-            "created": 3,
-        }
-    )
-    table.put_item(
-        Item={
-            "forum_name": "the-key",
-            "subject": "456",
-            "username": "johndoe",
-            "created": 1,
-        }
-    )
-    table.put_item(
-        Item={
-            "forum_name": "the-key",
-            "subject": "789",
-            "username": "johndoe",
-            "created": 2,
-        }
-    )
-    table.put_item(
-        Item={
-            "forum_name": "the-key",
-            "subject": "159",
-            "username": "janedoe",
-            "created": 2,
-        }
-    )
-    table.put_item(
-        Item={
-            "forum_name": "the-key",
-            "subject": "601",
-            "username": "janedoe",
-            "created": 5,
-        }
-    )
-
-    # Test a query returning all johndoe items
-    results = table.query(
-        KeyConditionExpression=Key("username").eq("johndoe") & Key("created").gt(0),
-        ScanIndexForward=True,
-        IndexName="TestGSI",
-    )
-    assert results["ScannedCount"] == 3
-    expected = ["456", "789", "123"]
-    for index, item in enumerate(results["Items"]):
-        assert item["subject"] == expected[index]
-
-    # Return all johndoe items again, but in reverse
-    results = table.query(
-        KeyConditionExpression=Key("username").eq("johndoe") & Key("created").gt(0),
-        ScanIndexForward=False,
-        IndexName="TestGSI",
-    )
-    for index, item in enumerate(reversed(results["Items"])):
-        assert item["subject"] == expected[index]
-
-    # Filter the creation to only return some of the results
-    # And reverse order of hash + range key
-    results = table.query(
-        KeyConditionExpression=Key("created").gt(1) & Key("username").eq("johndoe"),
-        IndexName="TestGSI",
-    )
-    assert results["Count"] == 2
-
-    # Filter to return no results
-    results = table.query(
-        KeyConditionExpression=Key("username").eq("janedoe") & Key("created").gt(9),
-        IndexName="TestGSI",
-    )
-    assert results["Count"] == 0
-
-    results = table.query(
-        KeyConditionExpression=Key("username").eq("janedoe") & Key("created").eq(5),
-        IndexName="TestGSI",
-    )
-    assert results["Count"] == 1
-
-    # Test range key sorting
-    results = table.query(
-        KeyConditionExpression=Key("username").eq("johndoe") & Key("created").gt(0),
-        IndexName="TestGSI",
-    )
-    expected = [Decimal("1"), Decimal("2"), Decimal("3")]
-    for index, item in enumerate(results["Items"]):
-        assert item["created"] == expected[index]
-
-
-@mock_aws
 def test_boto3_update_table_throughput():
     dynamodb = boto3.resource("dynamodb", region_name="us-east-1")
 
     # Create the DynamoDB table.
-    table = dynamodb.create_table(
+    dynamodb.create_table(
         TableName="users",
         KeySchema=[
             {"AttributeName": "forum_name", "KeyType": "HASH"},
@@ -1060,234 +890,6 @@ def test_update_table_gsi_throughput():
 
     table = dynamodb.Table("users")
     assert len(table.global_secondary_indexes) == 0
-
-
-@mock_aws
-def test_query_pagination():
-    table = _create_table_with_range_key()
-    for i in range(10):
-        table.put_item(
-            Item={
-                "forum_name": "the-key",
-                "subject": f"{i}",
-                "username": "johndoe",
-                "created": Decimal("3"),
-            }
-        )
-
-    page1 = table.query(KeyConditionExpression=Key("forum_name").eq("the-key"), Limit=6)
-    assert page1["Count"] == 6
-    assert page1["ScannedCount"] == 6
-    assert len(page1["Items"]) == 6
-
-    page2 = table.query(
-        KeyConditionExpression=Key("forum_name").eq("the-key"),
-        Limit=6,
-        ExclusiveStartKey=page1["LastEvaluatedKey"],
-    )
-    assert page2["Count"] == 4
-    assert page2["ScannedCount"] == 4
-    assert len(page2["Items"]) == 4
-    assert "LastEvaluatedKey" not in page2
-
-    results = page1["Items"] + page2["Items"]
-    subjects = set([int(r["subject"]) for r in results])
-    assert subjects == set(range(10))
-
-
-@mock_aws
-def test_query_by_local_secondary_index():
-    dynamodb = boto3.resource("dynamodb", region_name="us-east-1")
-
-    table = dynamodb.create_table(
-        TableName="test",
-        KeySchema=[
-            {"AttributeName": "id", "KeyType": "HASH"},
-            {"AttributeName": "range_key", "KeyType": "RANGE"},
-        ],
-        AttributeDefinitions=[
-            {"AttributeName": "id", "AttributeType": "S"},
-            {"AttributeName": "range_key", "AttributeType": "S"},
-            {"AttributeName": "lsi_range_key", "AttributeType": "S"},
-        ],
-        ProvisionedThroughput={"ReadCapacityUnits": 1, "WriteCapacityUnits": 1},
-        LocalSecondaryIndexes=[
-            {
-                "IndexName": "test_lsi",
-                "KeySchema": [
-                    {"AttributeName": "id", "KeyType": "HASH"},
-                    {"AttributeName": "lsi_range_key", "KeyType": "RANGE"},
-                ],
-                "Projection": {"ProjectionType": "ALL"},
-            }
-        ],
-    )
-
-    table.put_item(
-        Item={
-            "id": "1",
-            "range_key": "1",
-            "col1": "val1",
-            "lsi_range_key": "1",
-        },
-    )
-
-    table.put_item(
-        Item={
-            "id": "1",
-            "range_key": "2",
-            "col1": "val2",
-            "lsi_range_key": "2",
-        },
-    )
-
-    table.put_item(
-        Item={"id": "3", "range_key": "1", "col1": "val3"},
-    )
-
-    res = table.query(
-        KeyConditionExpression=Key("id").eq("1") & Key("lsi_range_key").eq("1"),
-        IndexName="test_lsi",
-    )
-    assert res["Count"] == 1
-    assert res["Items"] == [
-        {"id": "1", "range_key": "1", "col1": "val1", "lsi_range_key": "1"}
-    ]
-
-    res = table.query(
-        KeyConditionExpression=Key("id").eq("1") & Key("lsi_range_key").eq("2"),
-        IndexName="test_lsi",
-        ConsistentRead=True,
-    )
-    assert res["Count"] == 1
-    assert res["Items"] == [
-        {"id": "1", "range_key": "2", "col1": "val2", "lsi_range_key": "2"}
-    ]
-
-
-@mock_aws
-def test_scan_by_index():
-    dynamodb = boto3.client("dynamodb", region_name="us-east-1")
-
-    dynamodb.create_table(
-        TableName="test",
-        KeySchema=[
-            {"AttributeName": "id", "KeyType": "HASH"},
-            {"AttributeName": "range_key", "KeyType": "RANGE"},
-        ],
-        AttributeDefinitions=[
-            {"AttributeName": "id", "AttributeType": "S"},
-            {"AttributeName": "range_key", "AttributeType": "S"},
-            {"AttributeName": "gsi_col", "AttributeType": "S"},
-            {"AttributeName": "gsi_range_key", "AttributeType": "S"},
-            {"AttributeName": "lsi_range_key", "AttributeType": "S"},
-        ],
-        ProvisionedThroughput={"ReadCapacityUnits": 1, "WriteCapacityUnits": 1},
-        GlobalSecondaryIndexes=[
-            {
-                "IndexName": "test_gsi",
-                "KeySchema": [
-                    {"AttributeName": "gsi_col", "KeyType": "HASH"},
-                    {"AttributeName": "gsi_range_key", "KeyType": "RANGE"},
-                ],
-                "Projection": {"ProjectionType": "ALL"},
-                "ProvisionedThroughput": {
-                    "ReadCapacityUnits": 1,
-                    "WriteCapacityUnits": 1,
-                },
-            }
-        ],
-        LocalSecondaryIndexes=[
-            {
-                "IndexName": "test_lsi",
-                "KeySchema": [
-                    {"AttributeName": "id", "KeyType": "HASH"},
-                    {"AttributeName": "lsi_range_key", "KeyType": "RANGE"},
-                ],
-                "Projection": {"ProjectionType": "ALL"},
-            }
-        ],
-    )
-
-    dynamodb.put_item(
-        TableName="test",
-        Item={
-            "id": {"S": "1"},
-            "range_key": {"S": "1"},
-            "col1": {"S": "val1"},
-            "gsi_col": {"S": "1"},
-            "gsi_range_key": {"S": "1"},
-            "lsi_range_key": {"S": "1"},
-        },
-    )
-
-    dynamodb.put_item(
-        TableName="test",
-        Item={
-            "id": {"S": "1"},
-            "range_key": {"S": "2"},
-            "col1": {"S": "val2"},
-            "gsi_col": {"S": "1"},
-            "gsi_range_key": {"S": "2"},
-            "lsi_range_key": {"S": "2"},
-        },
-    )
-
-    dynamodb.put_item(
-        TableName="test",
-        Item={"id": {"S": "3"}, "range_key": {"S": "1"}, "col1": {"S": "val3"}},
-    )
-
-    res = dynamodb.scan(TableName="test")
-    assert res["Count"] == 3
-    assert len(res["Items"]) == 3
-
-    res = dynamodb.scan(TableName="test", Limit=1)
-    assert res["Count"] == 1
-    assert res["ScannedCount"] == 1
-
-    res = dynamodb.scan(TableName="test", ExclusiveStartKey=res["LastEvaluatedKey"])
-    assert res["Count"] == 2
-    assert res["ScannedCount"] == 2
-
-    res = dynamodb.scan(TableName="test", IndexName="test_gsi")
-    assert res["Count"] == 2
-    assert res["ScannedCount"] == 2
-    assert len(res["Items"]) == 2
-
-    res = dynamodb.scan(TableName="test", IndexName="test_gsi", Limit=1)
-    assert res["Count"] == 1
-    assert res["ScannedCount"] == 1
-    assert len(res["Items"]) == 1
-    last_eval_key = res["LastEvaluatedKey"]
-    assert last_eval_key["id"]["S"] == "1"
-    assert last_eval_key["gsi_col"]["S"] == "1"
-    assert last_eval_key["gsi_range_key"]["S"] == "1"
-
-    res = dynamodb.scan(
-        TableName="test", IndexName="test_gsi", ExclusiveStartKey=last_eval_key
-    )
-    assert res["Count"] == 1
-    assert res["ScannedCount"] == 1
-
-    res = dynamodb.scan(TableName="test", IndexName="test_lsi")
-    assert res["Count"] == 2
-    assert res["ScannedCount"] == 2
-    assert len(res["Items"]) == 2
-
-    res = dynamodb.scan(TableName="test", IndexName="test_lsi", ConsistentRead=True)
-    assert res["Count"] == 2
-    assert res["ScannedCount"] == 2
-    assert len(res["Items"]) == 2
-
-    res = dynamodb.scan(TableName="test", IndexName="test_lsi", Limit=1)
-    assert res["Count"] == 1
-    assert res["ScannedCount"] == 1
-    assert len(res["Items"]) == 1
-    last_eval_key = res["LastEvaluatedKey"]
-    assert last_eval_key["id"]["S"] == "1"
-    assert last_eval_key["range_key"]["S"] == "1"
-    assert last_eval_key["lsi_range_key"]["S"] == "1"
 
 
 @mock_aws
